refactor: simplify asset portfolio component
